+++ src/main/java/com/example/barber/repositories/EmployeeInfoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeInfoRepository extends JpaRepository<EmployeeInfo, Integer> {
+public interface EmployeeInfoRepository extends JpaRepository<EmployeeInfo, Long> {
 }
