+++ src/main/java/com/example/barber/repositories/PriceListRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PriceListRepository extends JpaRepository<PriceList, Integer> {
+public interface PriceListRepository extends JpaRepository<PriceList, Long> {
 }
